Rework the example/ structure
