+++ example/db/assets/structure.sql
