+++ example/db/index.js
@@ -1,0 +1,2 @@
+export { User } from './user.js';
+export { Login } from './login.js';
